move histogramming functions in histogramming.py
