+++ plothist/plotters.py
@@ -7,7 +7,6 @@ import numpy as np
 import boost_histogram as bh
 import matplotlib.pyplot as plt
 from matplotlib.transforms import Bbox
-import warnings
 import re
 from plothist.comparison import (
     get_comparison,
@@ -16,6 +15,7 @@ from plothist.comparison import (
     _check_uncertainty_type,
     _is_unweighted,
 )
+from plothist.histogramming import _flatten_2d_hist
 from plothist.plothist_style import set_fitting_ylabel_fontsize
 
 
@@ -39,6 +39,7 @@ def create_comparison_figure(
     hspace : float, optional
         Height spacing between subplots. Default is 0.15.
 
+
     Returns
     -------
     fig : matplotlib.figure.Figure
@@ -58,190 +59,6 @@ def create_comparison_figure(
         _ = ax.xaxis.set_ticklabels([])
 
     return fig, axes
-
-
-def create_axis(data, bins, range=None):
-    """
-    Create an axis object for histogram binning based on the input data and parameters.
-
-    Parameters
-    ----------
-    data : array-like
-        The input data for determining the axis range.
-    bins : int or array-like
-        The number of bins or bin edges for the axis.
-    range : None or tuple, optional
-        The range of the axis. If None, it will be determined based on the data.
-
-    Returns
-    -------
-    Axis object
-        An axis object for histogram binning.
-
-    Raises
-    ------
-    ValueError
-        If the range parameter is invalid or not finite.
-    """
-
-    try:
-        N = len(bins)
-    except TypeError:
-        N = 1
-
-    if N > 1:
-        if range is not None:
-            warnings.warn(f"Custom binning -> ignore supplied range ({range}).")
-        return bh.axis.Variable(bins)
-
-    # Inspired from np.histograms
-    if range is not None:
-        x_min = min(data) if range[0] == "min" else range[0]
-        x_max = max(data) if range[1] == "max" else range[1]
-        if x_min > x_max:
-            raise ValueError(
-                f"Range of [{x_min}, {x_max}] is not valid. Max must be larger than min."
-            )
-        if not (np.isfinite(x_min) and np.isfinite(x_max)):
-            raise ValueError(f"Range of [{x_min}, {x_max}] is not finite.")
-    elif data.size == 0:
-        # handle empty arrays. Can't determine range, so use 0-1.
-        x_min, x_max = 0, 1
-    else:
-        x_min, x_max = min(data), max(data)
-        if not (np.isfinite(x_min) and np.isfinite(x_max)):
-            raise ValueError(f"Autodetected range of [{x_min}, {x_max}] is not finite.")
-
-    # expand empty range to avoid divide by zero
-    if x_min == x_max:
-        x_min = x_min - 0.5
-        x_max = x_max + 0.5
-
-    return bh.axis.Regular(bins, x_min, x_max)
-
-
-def _flatten_2d_hist(hist):
-    # TODO: support other storages, generalise to N dimensions
-    """
-    Flatten a 2D histogram into a 1D histogram.
-
-    Parameters
-    ----------
-    hist : Histogram object
-        The 2D histogram to be flattened.
-
-    Returns
-    -------
-    Histogram object
-        The flattened 1D histogram.
-    """
-    n_bins = hist.axes[0].size * hist.axes[1].size
-    flatten_hist = bh.Histogram(
-        bh.axis.Regular(n_bins, 0, n_bins), storage=bh.storage.Weight()
-    )
-    flatten_hist[:] = np.c_[hist.values().flatten(), hist.variances().flatten()]
-    return flatten_hist
-
-
-def make_hist(data, bins=50, range=None, weights=1):
-    """
-    Create a histogram object and fill it with the provided data.
-
-    Parameters
-    ----------
-    data : array-like
-        1D array-like data used to fill the histogram.
-    bins : int or tuple, optional
-        Binning specification for the histogram (default is 50).
-        If an integer, it represents the number of bins.
-        If a tuple, it should be the explicit list of all bin edges.
-    range : tuple, optional
-        The range of values to consider for the histogram bins (default is None).
-        If None, the range is determined from the data.
-    weights : float or array-like, optional
-        Weight(s) to apply to the data points (default is 1).
-        If a float, a single weight is applied to all data points.
-        If an array-like, weights are applied element-wise.
-
-    Returns
-    -------
-    histogram : boost_histogram.Histogram
-        The filled histogram object.
-    """
-
-    axis = create_axis(data, bins, range)
-
-    if weights is None:
-        storage = bh.storage.Double()
-    else:
-        storage = bh.storage.Weight()
-
-    h = bh.Histogram(axis, storage=storage)
-    h.fill(data, weight=weights, threads=0)
-
-    # Check what proportion of the data is in the underflow and overflow bins
-    range_coverage = h.values().sum() / h.values(flow=True).sum()
-    # Issue a warning in more than 1% of the data is outside of the binning range
-    if range_coverage < 0.99:
-        warnings.warn(
-            f"Only {100*range_coverage:.2f}% of data contained in the binning range ({axis.edges[0]}, {axis.edges[-1]})."
-        )
-
-    return h
-
-
-def make_2d_hist(data, bins=(10, 10), range=(None, None), weights=1):
-    """
-    Create a 2D histogram object and fill it with the provided data.
-
-    Parameters
-    ----------
-    data : array-like
-        2D array-like data used to fill the histogram.
-    bins : tuple, optional
-        Binning specification for each dimension of the histogram (default is (10, 10)).
-        Each element of the tuple represents the number of bins for the corresponding dimension.
-        Also support explicit bin edges specification (for non-constant bin size).
-    range : tuple, optional
-        The range of values to consider for each dimension of the histogram (default is (None, None)).
-        If None, the range is determined from the data for that dimension.
-        The tuple should have the same length as the data.
-    weights : float or array-like, optional
-        Weight(s) to apply to the data points (default is 1).
-        If a float, a single weight is applied to all data points.
-        If an array-like, weights are applied element-wise.
-
-    Returns
-    -------
-    histogram : boost_histogram.Histogram
-        The filled 2D histogram object.
-
-    Raises
-    ------
-    ValueError
-        If the data does not have two components or if the lengths of x and y are not equal.
-    """
-    if len(data) != 2:
-        raise ValueError("data should have two components, x and y")
-    if len(data[0]) != len(data[1]):
-        raise ValueError("x and y must have the same length.")
-
-    h = bh.Histogram(
-        create_axis(data[0], bins[0], range[0]),
-        create_axis(data[1], bins[1], range[1]),
-        storage=bh.storage.Weight(),
-    )
-    h.fill(*data, weight=weights, threads=0)
-
-    # Check what proportion of the data is in the underflow and overflow bins
-    range_coverage = h.values().sum() / h.values(flow=True).sum()
-    # Issue a warning in more than 1% of the data is outside of the binning range
-    if range_coverage < 0.99:
-        warnings.warn(
-            f"Only {100*range_coverage:.2f}% of data contained in the binning range."
-        )
-
-    return h
 
 
 def plot_hist(hist, ax, **kwargs):
@@ -367,59 +184,6 @@ def plot_function(func, range, ax, stacked=False, npoints=1000, **kwargs):
             [f(x) for f in func],
             **kwargs,
         )
-
-
-def _make_hist_from_function(func, ref_hist):
-    """
-    Create a histogram from a function and a reference histogram.
-    The returned histogram has the same binning as the reference histogram and
-    is filled with the function evaluated at the bin centers of the reference histogram.
-
-    Parameters
-    ----------
-    func : function
-        1D function. The function should support vectorization (i.e. accept a numpy array as input).
-    ref_hist : boost_histogram.Histogram
-        The reference 1D histogram to use for the binning.
-
-    Returns
-    -------
-    hist : boost_histogram.Histogram
-        The histogram filled with the function.
-
-    Raises
-    ------
-    ValueError
-        If the reference histogram is not 1D.
-    """
-    if len(ref_hist.axes) != 1:
-        raise ValueError("The reference histogram must be 1D.")
-
-    hist = bh.Histogram(ref_hist.axes[0], storage=bh.storage.Weight())
-    hist[:] = np.c_[
-        func(ref_hist.axes[0].centers), np.zeros_like(ref_hist.axes[0].centers)
-    ]
-    return hist
-
-
-def _check_hist_or_func_model(components):
-    """
-    Check whether all components of a model are either all histograms or all functions.
-
-    Parameters
-    ----------
-    components : list
-        The list of model components.
-
-    Raises
-    ------
-    ValueError
-        If the model components are not all histograms or all functions.
-    """
-    if not all(isinstance(x, bh.Histogram) for x in components) and not all(
-        callable(x) for x in components
-    ):
-        raise ValueError("All model components must be either histograms or functions.")
 
 
 def plot_2d_hist_with_projections(
@@ -921,6 +685,26 @@ def _get_math_text(text):
         return match.group(1)
     else:
         return text
+
+
+def _check_hist_or_func_model(components):
+    """
+    Check whether all components of a model are either all histograms or all functions.
+
+    Parameters
+    ----------
+    components : list
+        The list of model components.
+
+    Raises
+    ------
+    ValueError
+        If the model components are not all histograms or all functions.
+    """
+    if not all(isinstance(x, bh.Histogram) for x in components) and not all(
+        callable(x) for x in components
+    ):
+        raise ValueError("All model components must be either histograms or functions.")
 
 
 def plot_mc(
